api: behaviour: use Accept header to control watch mode
Rather than using a `?watch=true` query parameter as was originally
proposed in the RFC, use the `Accept` header to negotiate the
`Content-Type`.

The motiviation for this change was originally in discovering that the
server did not serve events properly when compression was requested
clientside (as it is when using the javascript `EventSource`) since the
minimum gzip size was not reached. I found that the `BuildEvents`
endpoint sets the minimum gzip size to 0 as a special case in a
compression wrappa. Driving out SSE for `ListAllJobs` did not make this
at all obvious. While I could have added another special case for
`ListAllJobs`, I envision many endpoints following this pattern - so,
instead, wrap every handler in a handler that sets the minimum gzip size
to 0 if the `Accept` header is `text/event-stream`, and the default
otherwise. This way, if/when we implement more watch endpoints, we won't
run into this issue again.

While this wrapper could check for `?watch=true`, I realized that
`Accept` is the canonical way to request data be served in a particular
way.

P.S. I intentionally didn't wrap the `BuildEvents` handler in the
wrapper, since there may be other consumers of the `BuildEvents`
endpoint that aren't setting `Accept: text/event-stream`, and I don't
want to break them.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>